Update item selectors to 1.20.5
From my testing, this commit fixes item modifiers, the ground creation
animation and item drops.

Mojang seems to have added a failsafe to this update, updating existing
items with "Count" and "tag" data to the associated new format.
Therefore, not all instances of these are syntactically correct in this
code; only the ones in selectors. (Trident still defaults to "Count", so
this made more sense for my sanity.)